moved constants.dart to core
