+++ lib/models/app_state.dart
@@ -6,13 +6,13 @@ import 'package:flutter/services.dart' show rootBundle;
 import 'package:http/http.dart';
 import 'package:path_provider/path_provider.dart';
 import 'package:shared_preferences/shared_preferences.dart';
-import 'constants.dart';
+import '../src/core/utils/constants.dart';
 import 'dictionary.dart';
 import 'culture.dart';
 import 'language.dart';
 
 export 'dictionary.dart';
-export 'constants.dart';
+export '../src/core/utils/constants.dart';
 export 'culture.dart';
 export 'language.dart';
 export 'user_account.dart';
